requirements.txt
asgiref==3.8.1
contourpy==1.3.1 
cycler==0.12.1 
Django==5.1.7 
django-filter==25.1 
djangorestframework==3.15.2 
fonttools==4.56.0 
kiwisolver==1.4.8 
matplotlib==3.10.1 
numpy==2.2.3 
packaging==24.2 
pandas==2.2.3 
pillow==11.1.0 
pyparsing==3.2.1 
python-dateutil==2.9.0.post0 
python-dotenv==1.1.0 
pytz==2025.1 six==1.17.0 sqlparse==0.5.3 tzdata==2025.1 
fpdf
python-docx

